change in  get method

+++ api/views.py
@@ -24,27 +24,34 @@ def api_home_page(request):
             return Response(serializeData.data, status=status.HTTP_201_CREATED)
         return Response(serializeData.errors, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['PUT', 'DELETE'])
+@api_view(['PUT', 'DELETE', 'GET'])
 def book_detail(request, pk):
     try:
         book = Book.objects.get(pk=pk)
     except Book.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
     
+    if request.method == 'GET':
+        
+        serializeData = BookSerializer(book)
+        return Response(serializeData.data, status=status.HTTP_200_OK)
+    
     if request.method == 'DELETE':
         
         book.delete()
         data = request.data
         serializeData = BookSerializer(book)
-        return Response(serializeData.data, status=status.HTTP_204_NO_CONTENT)
+        return Response("Content Successfully Deleted!")
     
-    elif request.method == 'PUT':
+    if request.method == 'PUT':
+        
         data = request.data
         serializeData = BookSerializer(book, data=data)
         if serializeData.is_valid():
             serializeData.save()
             return Response(serializeData.data)
         return Response(serializeData.errors, status=status.HTTP_400_BAD_REQUEST)
+ 
     
     
 class CreateUserView(generics.CreateAPIView):
